Entrance enhancements
More options in the option to freeze the position of Monkey Madness (good for Specter 1 goals)
Also re-ordered the options to group similar options together while I was there, and standardized the formatting of option descriptions.
Fixed a bug where quickly entering a level would send the player to the level at the first entrance.
Also fixed a logic error reported on Discord.