Datastorage implementation
Implemented Datastorage to keep some values permanently and use them later.
- Changed CrC Button vanilla behavior to prepare for Room Randomization:
if the button gets activated it stores a value in Datastorage permanently. If you are in the right room it then tells the client to unlock the CrC Boss Room door by setting both the Visual and Hitbox to 0,enabling the transition.