Change to coin display behavior in Level Select
Coins information is not displayed anymore in the Level Selection menu.
They are moved into a temporary table when in Level Select and then moved back when the player enters a Level.
This means they will never show/detect on the Level Select but will still be able to be tracked when playing.