- Monkey count check when exiting a level
(First part of the syncprogress correction)

Made a function, linked with a table, that gets all the monkeys caught status for the last entered level and compare it to the RAM address value for the monkey counts,then replace if needed.

Made this change because that PPM Vanilla unlocks still base itself on monkeyCounts values and they could somtimes be wrong.