Fixed a vanilla mechanic!
WSW Gong Room mechanics does not get triggered fully when entering from Middle Room door.

The trigger is in 2 parts : Being close to Gong_Entry door, then stepping far enough near the gong.

We bypass it by setting the roomState to 1 when camera is locked and we are in the right room (That means we are probably near the gong)